[brain] op_profiler.hpp include surgery: decouple JSON from device_operation hot path
Moves the heavy JSON/Tracy profiling code out of op_profiler.hpp and into
a new op_profiler_json.cpp, compiled once. device_operation.hpp now includes
the lightweight op_profiler_serialize.hpp instead of the full op_profiler.hpp,
removing nlohmann/json.hpp + tt_metal.hpp from the 1000+ TU include chain.

Also incorporates two post-branch fixes:
- #40499: deduplicate kernels by source in get_kernels_json to avoid blowing
  past Tracy's 64 KiB message limit in fused ops with many kernels
- #40667: three-tier message truncation (pretty → compact → drop kernel_info)
  with proper warning on final overflow, fixing CSV quotechar corruption

Addresses review comment: update copyright to © 2026 Tenstorrent AI ULC
on the two new files.
